# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ProofU API
  version: 0.1.0
  description: |
    Career source of truth: evidence-linked career data, job posting analysis, explainable
    matching, traceable application documents, immutable submission snapshots and reviews.

    Conventions (docs/api/conventions.md):
    * Base path `/api/v1`, JSON, UTF-8.
    * `201` on create, `202` + `jobId` for asynchronous work, polled via `GET /jobs/{id}`.
    * Errors are RFC 9457 Problem Details with a stable `code` property.
    * Cursor pagination (`cursor`, `limit` ≤ 100). Mutations carry `version`/`revision` for optimistic locking.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security:
  - session: []
...
paths:
...
  /job-posting-snapshots/{id}/analysis-jobs:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [jobs, async]
      operationId: startPostingAnalysis
      summary: Extract requirements from a snapshot (asynchronous)
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        default:
          $ref: "#/components/responses/Problem"
...
  /applications/{id}/documents:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [documents]
      operationId: listApplicationDocuments
      summary: Documents of an application, newest first
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [documents]
      operationId: createDocument
      summary: Create an empty document for an application
      description: The document has no version until one is generated or saved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, title]
              properties:
                type:
                  $ref: "#/components/schemas/DocumentType"
                title:
                  type: string
                  maxLength: 200
                language:
                  type: string
                  default: ko
      responses:
        "201":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        default:
          $ref: "#/components/responses/Problem"
  /documents/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [documents]
      operationId: getDocument
      summary: Document with its latest version and the template sections
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        default:
          $ref: "#/components/responses/Problem"
  /documents/{id}/generation-jobs:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [documents, async]
      operationId: startDocumentGeneration
      summary: Generate a new AI-drafted version from selected sources (asynchronous)
      description: |
        Sources default to the application's approved requirements and the match candidates the
        user ACCEPTED (F05); `claimIds`/`requirementIds` narrow that set and never widen it.
        The model only phrases and arranges those facts. The server whitelists every reference,
        caps each block's certainty at the evidence level of the claims it cites, and stores the
        result as an unapproved AI version. Job type `document.generation`; the job result carries
        `versionId`. Rejected with 422 when nothing has been accepted for the application.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                templateVersion:
                  type: string
                  default: ko-v1
                parentVersionId:
                  $ref: "#/components/schemas/Uuid"
                claimIds:
                  type: array
                  items:
                    $ref: "#/components/schemas/Uuid"
                requirementIds:
                  type: array
                  items:
                    $ref: "#/components/schemas/Uuid"
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        default:
          $ref: "#/components/responses/Problem"
  /documents/{id}/versions:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [documents]
      operationId: listDocumentVersions
      summary: List versions of a document
      responses:
        "200":
          description: Versions, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DocumentVersion"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [documents]
      operationId: createDocumentVersion
      summary: Save a user-edited version
      description: |
        Users may edit text, drop or reorder blocks and approve INFERRED/UNSUPPORTED ones. A block
        may not cite a source its parent block did not, and its certainty can only go down; blocks
        typed by hand are UNSUPPORTED. Violations are 422 DOMAIN_RULE_VIOLATION. The parent must be
        the document's latest version (409 CONFLICT_STALE_VERSION otherwise).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blocks]
              properties:
                parentVersionId:
                  description: The document's latest version; omitted only while the document has none
                  $ref: "#/components/schemas/Uuid"
                label:
                  type: string
                blocks:
                  type: array
                  items:
                    $ref: "#/components/schemas/DocumentBlock"
      responses:
        "201":
          description: Version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersion"
        default:
          $ref: "#/components/responses/Problem"
  /document-versions/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [documents]
      operationId: getDocumentVersion
      summary: Get a document version with its blocks and provenance
      responses:
        "200":
          description: Version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersion"
        default:
          $ref: "#/components/responses/Problem"
...
  /jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [async]
      operationId: getJob
      summary: Poll an asynchronous job
      responses:
        "200":
          description: Job status and, when finished, its result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC-issued session token. Provider is TBD (docs/project/open-decisions.md).
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Accepted:
      description: Work accepted; poll `GET /jobs/{jobId}`
      content:
        application/json:
          schema:
            type: object
            required: [jobId]
            properties:
              jobId:
                $ref: "#/components/schemas/Uuid"
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ProblemDetail:
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          $ref: "#/components/schemas/ErrorCode"
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
              rejectedValue: {}
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - REAUTHENTICATION_REQUIRED
        - FORBIDDEN_WORKSPACE
        - NOT_FOUND
        - CONFLICT_STALE_VERSION
        - INVALID_STATUS_TRANSITION
        - SNAPSHOT_IMMUTABLE
        - DOMAIN_RULE_VIOLATION
        - UNSUPPORTED_CLAIM_IN_EXPORT
        - RATE_LIMITED
        - AI_PROVIDER_UNAVAILABLE
        - AI_REFUSED
        - AI_OUTPUT_INVALID
        - AI_BUDGET_EXCEEDED
        - NOT_IMPLEMENTED
        - INTERNAL_ERROR
...
    Certainty:
      type: string
      enum: [SUPPORTED, INFERRED, UNSUPPORTED]
...
    JobStatus:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
...
    DocumentType:
      type: string
      enum: [RESUME, COVER_LETTER, PORTFOLIO]
    Document:
      type: object
      required: [id, applicationId, type, title, language, version, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        applicationId:
          $ref: "#/components/schemas/Uuid"
        type:
          $ref: "#/components/schemas/DocumentType"
        title:
          type: string
        language:
          type: string
        latestVersionId:
          $ref: "#/components/schemas/Uuid"
        latestVersionCreatedBy:
          type: string
          enum: [USER, AI]
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/Document"
        - type: object
          required: [sections, pendingApprovalCount]
          properties:
            latestVersion:
              $ref: "#/components/schemas/DocumentVersion"
            sections:
              description: Sections of the latest version's template (or the current template when there is no version)
              type: array
              items:
                $ref: "#/components/schemas/TemplateSection"
            pendingApprovalCount:
              description: Blocks of the latest version that block an export until approved
              type: integer
    TemplateSection:
      type: object
      required: [id, title, guidance]
      properties:
        id:
          type: string
        title:
          type: string
        guidance:
          type: string
    DocumentBlock:
      type: object
      required: [blockId, text, certainty]
      properties:
        blockId:
          description: "`<section id>-<n>`; stable across versions"
          type: string
        text:
          type: string
        claimRefs:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
        evidenceRefs:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
        requirementRefs:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
        certainty:
          $ref: "#/components/schemas/Certainty"
        warnings:
          type: array
          items:
            type: string
        approvedByUser:
          type: boolean
          default: false
    DocumentVersion:
      type: object
      required: [id, documentId, blocks, templateVersion, createdBy, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        documentId:
          $ref: "#/components/schemas/Uuid"
        parentId:
          $ref: "#/components/schemas/Uuid"
        label:
          type: string
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/DocumentBlock"
        templateVersion:
          type: string
        promptVersion:
          type: string
        modelRef:
          type: string
        createdBy:
          type: string
          enum: [USER, AI]
        createdAt:
          type: string
          format: date-time
        provenance:
          description: Source revisions each block was derived from; never changes after the version is written
          type: array
          items:
            $ref: "#/components/schemas/ProvenanceLink"
    ProvenanceLink:
      type: object
      required: [blockId, sourceType, sourceId, sourceRevision, relation]
      properties:
        blockId:
          type: string
        sourceType:
          type: string
          enum: [CAREER_ENTRY, PROJECT, ACHIEVEMENT, CLAIM, EVIDENCE, REQUIREMENT]
        sourceId:
          $ref: "#/components/schemas/Uuid"
        sourceRevision:
          type: integer
          format: int64
        relation:
          type: string
          enum: [DERIVED_FROM, CITES, ADDRESSES]
    Job:
      type: object
      required: [id, type, status, attempts, createdAt]
      description: |
        Poll until status is SUCCEEDED, FAILED or CANCELLED. errorCode is one of the API
        ErrorCode values (e.g. AI_BUDGET_EXCEEDED, AI_REFUSED, AI_OUTPUT_INVALID,
        AI_PROVIDER_UNAVAILABLE) or a job-specific code.
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        type:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        attempts:
          type: integer
        errorCode:
          type: string
        result:
          description: Job-type specific result, present when SUCCEEDED
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time